+++ common/ui-compose/src/commonMain/kotlin/dev/t1r/themebuilder/ui/compose/RootUi.kt
@@ -12,6 +12,7 @@ import dev.t1r.themebuilder.feature.materialcolorspalette.MaterialColorsPaletteC
 import dev.t1r.themebuilder.feature.root.RootComponent
 import dev.t1r.themebuilder.feature.root.RootComponent.Child
 import dev.t1r.themebuilder.feature.root.RootComponent.Model
+import dev.t1r.themebuilder.ui.compose.export.ExportContent
 
 @OptIn(ExperimentalDecomposeApi::class)
 @Composable
